+++ apiRouters/categoriesRouter.js
@@ -1,0 +1,6 @@
+const categoriesRouter = require("express").Router();
+const { getCategories } = require("../controllers/categories.controller");
+
+categoriesRouter.get("/", getCategories);
+
+module.exports = categoriesRouter;
