+++ apiRouters/usersRouter.js
@@ -1,0 +1,6 @@
+const usersRouter = require("express").Router();
+const { getUsers } = require("../controllers/users.controller");
+
+usersRouter.get("/", getUsers);
+
+module.exports = usersRouter;
